Put the Response and RawResponse classes in their own file.Added a ResponsesHandler class
