+++ code/includes/Response.hpp
@@ -3,40 +3,7 @@
 
 # include <string>
 # include <map>
-# include <sys/types.h>
 # include <stdint.h>
-
-/**
- * @brief This class stores all the data that will be sent to the client as a
- * response, compacted in a single string and a fd for the body. This class take
- * responsability for the closing of the fd.
- * It moves the data from the file to the socket like a stream, without allocation.
- * This class is designed to be minimalistic.
- */
-class RawResponse
-{
-private:
-	/**
-	 * @brief A string containing the first part of the response, that means :
-	 * the status line, the headers, the blank line and, if there is a static one,
-	 * a body.
-	 */
-	std::string	_firstPart;
-	/**
-	 * @brief The number of character from _firstPart that need to be written
-	 * to the socket. It should be in range [-1, firstPart.size()] : -1 meaning
-	 * there was an error, 0 meaning there is nothing to read anymore.
-	 */
-	ssize_t		_numCharsToWrite;
-	/**
-	 * @brief The file descriptor of the body.If there is no body, or it has already
-	 * been included in firstPart, this variable is set to -1.
-	 */
-	int			_bodyFd;
-
-public:
-	ssize_t	writeToFd(int fd);
-};
 
 /**
  * @brief The classe that stores all the data that will be sent to the client.
